+++ src/extractor.js
@@ -1,6 +1,6 @@
 const path = require('path')
-const { writeJson, readFile, existsSync } = require('fs-extra')
-const glob = require('glob-promise')
+const { writeJsonSync, readFileSync, existsSync } = require('fs-extra')
+const glob = require('glob')
 const merge = require('deepmerge')
 
 /**
@@ -8,13 +8,13 @@ const merge = require('deepmerge')
  * @param {RegExp|string} match
  * @param {string} filesExt
  */
-async function jsonMatchingContent(currentDir, match, filesExt) {
+function jsonMatchingContent(currentDir, match, filesExt) {
     const regex = new RegExp(match, 'gm')
-    const files = await glob(`${currentDir}/**/*.{${filesExt}}`)
+    const files = glob.sync(`${currentDir}/**/*.{${filesExt}}`)
     const jsonMatchedContentObj = {}
 
     for (const file of files) {
-        const fileContent = await readFile(file)
+        const fileContent = readFileSync(file)
         let matchedContent
 
         while ((matchedContent = regex.exec(fileContent)) !== null) {
@@ -40,41 +40,39 @@ async function jsonMatchingContent(currentDir, match, filesExt) {
 /**
  * @param {string} locales
  * @param {import('./types').I18nExtractorOptions} options
- * @return {Promise}
+ * @return {string[]}
  */
-module.exports = async function (locales, options) {
+module.exports = function (locales, options) {
     const localesArr = locales.split(',')
 
     if (localesArr.length <= 0) {
-        return console.error('No locales were introduced as input, some are required.', 1)
+        throw Error('No locales were introduced as input, some are required.')
     }
 
-    let currentDir = options?.path || process.env.PWD
+    let currentDir = options?.path || process.cwd()
 
-    currentDir = path.isAbsolute(currentDir) ? currentDir : path.resolve(process.env.PWD, currentDir)
+    currentDir = path.isAbsolute(currentDir) ? currentDir : path.resolve(process.cwd(), currentDir)
 
     const generatedLocaleFilePathArr = []
 
     for (const locale of localesArr) {
-        generatedLocaleFilePathArr.push(path.resolve(currentDir, `${locale}.json`))
+        generatedLocaleFilePathArr.push(path.resolve(options.output, `${locale}.json`))
     }
 
-    let contentMatched = await jsonMatchingContent(currentDir, options.match, options.extensions)
+    const outputPaths = []
+    let contentMatched = jsonMatchingContent(currentDir, options.match, options.extensions)
     let filePath
 
+    // eslint-disable-next-line no-cond-assign
     while (filePath = generatedLocaleFilePathArr.pop()) {
         if (existsSync(filePath)) {
-            const originalFileContent = await readFile(filePath)
-    
-            contentMatched = merge(contentMatched, JSON.parse(originalFileContent))
+            contentMatched = merge(contentMatched, JSON.parse(readFileSync(filePath)))
         }
 
-        try {
-            await writeJson(filePath, contentMatched)
-        } catch (err) {
-            console.error('Something happened trying to write the locale file. Please try again.')
-        }
+        writeJsonSync(filePath, contentMatched, { encoding: 'utf8' })
+
+        outputPaths.push(filePath)
     }
 
-    console.log(`${locales} files generated successfully!`)
+    return outputPaths
 }