+++ src/plugin.js
@@ -1,9 +1,12 @@
+const File = require('laravel-mix/src/File')
 const path = require('path');
 const extractor = require('./extractor');
+const constants = require('./constants');
+const merge = require('deepmerge');
 
 class i18n {
-    /** @type {string} */
-    entry = '';
+    /** @type {File[]} */
+    entry = [];
 
     /** @type {import('./types').I18nMixOptions} */
     options = {};
@@ -18,20 +21,53 @@ class i18n {
     }
 
     /**
+     * Specifiy one or more dependencies that must
+     * be installed for this component to work
+     *
+     * @returns {import("laravel-mix/src/Dependencies").Dependency[]}
+     **/
+     dependencies() {
+        return ['@intlify/vue-i18n-loader'];
+    }
+
+    /**
      * Register the component.
      *
      * @param {string} entry
+     * @param {string} output
      * @param {import('./types').I18nMixOptions} options
      */
-    register(entry, options = {}) {
-        this.options = options;
+    register(entry, output, options = {}) {
+        this.options = merge.all([
+            {
+                extractor: {
+                    path: entry,
+                    output
+                }
+            },
+            options,
+            {
+                extractor: constants
+            }
+        ]);
 
-        this.options.extractor.path = entry;
+        if ('locales' in this.options.extractor) {
+            const entries = extractor(this.options.extractor.locales, this.options.extractor)
 
-        if ('extract' in this.options && this.options.extract) {
-            this.entry = extractor(this.options.extractor.locales, this.options?.extractor || {});
+            this.entry = entries.map(entry => new File(entry));
         }
     }
+
+    /**
+     * Assets to append to the webpack entry.
+     *
+     * @param {import('laravel-mix/src/builder/Entry')} entry
+     */
+    //  webpackEntry(entry) {
+    //     this.entry.forEach(file => {
+    //         entry.add(this. this.options.extractor.path, file, this.entry[0]);
+    //     });
+    // }
 
     /**
      * Rules to be merged with the underlying webpack rules.
@@ -39,13 +75,17 @@ class i18n {
      * @return {Array|Object}
      */
      webpackRules() {
+        if (this.entry.length === 0 || !this.options.loader) {
+            return [];
+        }
+
         return [
             {
                 test: /\.(json5?|ya?ml)$/,
                 type: 'javascript/auto',
                 loader: '@intlify/vue-i18n-loader',
                 include: [
-                    path.resolve(this.entry)
+                    path.resolve(this.options.extractor.output),
                 ]
             }
         ];